Bug#32365780: MYSQL SREVER CRASH - SEGMENTATION FAULT IN               QUERY_BLOCK::DECORRELATE_DERIVED_SCALAR_SUBQUERY_POST
A non-correlated scalar subquery transformed into a derived table,
if used in transforming another correlated subquery into a
derived table, results in a server exit.

The problem here is that, when the first transformation happens,
(ex: when a correlated predicate (of a scalar subquery) has another
scalar subquery as one of the operand), we replace all the instances
of the scalar subquery with a field in the derived table.
However this field is added as a hidden item to the query block
of the derived table during transformation.
Now, when the correlated predicate is looked at for
transformation, we have an operand (non-correlated field)
marked as hidden. However this field is added expecting
it to be visible in the derived table the gets created later.
Since it is marked as hidden, while creating the result
table, we skip the hidden items. This results in a missing
field in the derived table leading to server exit later.

Solution:
Fix by marking the field as not hidden once it is decided to
be added as a visible field to the derived table.

Reviewed By: Roy Lyseng <roy.lyseng@oracle.com>