Bug#32248705: PUTTING A MISSING ROW INTO JSON_OBJECT CAUSES JSON_EXTRACT TO FAIL
Some JSON functions return a non-NULL value even though one of their
arguments is NULL. But the value returned from not_null_tables()
suggests that they do return NULL on NULL input. This might make the
optimizer transform a LEFT JOIN into an INNER JOIN, even though the
INNER JOIN is not equivalent to the original LEFT JOIN.

The following functions are affected:

- JSON_ARRAY: If an argument is NULL, the function returns an array
which contains a JSON null literal. It does not return NULL.

- JSON_OBJECT: If a key argument is NULL, the function raises an
error. If a value argument is NULL, it returns an object which
contains a JSON null literal in one of the members.

- JSON_ARRAY_APPEND, JSON_ARRAY_INSERT, JSON_INSERT, JSON_REPLACE,
JSON_SET: These functions return NULL only if the first argument (the
JSON document) is NULL, or if one of the JSON path arguments is NULL.
If one of the values to insert into the document is NULL, a JSON null
literal is inserted at the given position, instead of returning NULL
from the function.

Fixed by adjusting the value returned by not_null_tables() for these
functions. For JSON_ARRAY and JSON_OBJECT, set null_on_null to false,
which makes Item_func do the right thing. For the rest of the
functions, update not_null_tables_cache in resolve_type() and
update_used_tables() and make it only contain tables that are
referenced from the first argument or any of the JSON path arguments.

Change-Id: If773227d9f5d78a759124aa5a5de3b7113c37f54