Bug#31578783: ASSERTION "TABLE_FUNCTION_JSON..PRINT"
A JSON_TABLE expression could in some cases be attempted written to
the optimizer trace before the temporary table backing the table
function had been created. This would fail while attempting to access
the Field objects of the non-existent temporary table.

It is fixed by printing "<column type not resolved yet>" instead of
the actual column type when the type information is not yet available.

Change-Id: I00668aae55ce8f637112c3e1e38f4e3894af8ae9