Bug #32423860 	REMOVE UNNECESSARY FULL FLUSH LIST SCAN IN IBT::TABLESPACE::CREATE
The full flush list scan in form of:
buf_LRU_flush_or_remove_pages(m_space_id, BUF_REMOVE_FLUSH_WRITE, nullptr);

is unnecessary. It was introduced as a fix for:
Bug #27636039 	CURRENT WORKING DIRECTORY (CWD) BECOMES INVALID AFTER THE
DAEMONIZE_OPT.TEST

but it was not because of the root cause, it was only masking the symptoms.
The real root cause was found and fixed in:

----
SHA-1: ed8e32fda3f04514a08c855b4cd7fee04a743944

* Bug#30678292 MAIN.DAEMONIZE_OPT MIGHT CAUSE OTHER TESTS TO FAIL

Test proved to be problematic, causing other tests to fail.

As test main.mysqld_daemon already covers functionality in --daemonize
option, this test can be removed.
----

As there was no problem in the innodb in the first place, the flush list scan
should be just removed.

RB#25810