BUG#32413284 FIXES A LOG MESSAGE IN AN AUTOTEST TEST
Modifies testUpgrade.cpp to fix the log which conveys the reason
why a test is being skipped.

WL#12876 introduces a new SYSFILE format for files on disk and
signals because of which non initial "DATA NODE" downgrades
might not be possible. So, the error message is fixed to say
the same.

Change-Id: I23fd14c09cfdc458e34967e752ffb03b37949552