Bug#32257395 AUTOTEST: TESTSRBANK FAILS IN AUTOTEST
bdih makes assumptions about that it transition through the different
NodeRecoveryStatus in an orderly fashion -> 1,2,3,4,5,6,, and so one.

For each state transition, handled by setNodeRecoveryStatus, it records
the current NdbTick for when each state was reached, and also call
check_node_recovery_state towards its end to ndbrequire that all 'ticks'
related to previous states are 'Valid'.

For the state 'INCLUDED_IN_HB_PROTOCOL' however, it is commented in the
code that it can transition directly to this state from both
NODE_NOT_RESTARTED_YET and NODE_FAILURE_COMPLETED, without first being in
the preceeding ALLOCATED_NODE_ID state. (also checked with a ndbrequire)

We did handle the direct transition from NODE_FAILURE_COMPLETED by also
setting the allocatedNodeIdTime 'tick' in such cases. However this check
missed that we could also transition from NODE_NOT_RESTARTED_YET state.

This patch add this state to the same if-branch which already
handled NODE_FAILURE_COMPLETED

Reviewed by: Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>