Bug#26086751: LIKE LOSES ESCAPE CLAUSE WHEN USED IN VIEW Bug#29224931: LIKE ... ESCAPE IS NOT WRITTEN TO DICTIONARY Bug#30211596: `EXPLAIN ... LIKE ... ESCAPE` DOESN'T OUTPUT `ESCAPE` CLAUSE IN THE WARNING
The code that prints a LIKE expression, does not print the optional
ESCAPE clause of the expression, even if it has been specified. This
causes issues like the ESCAPE clause missing in warnings and error
messages that refer to the LIKE expression, and also that the ESCAPE
clause is lost when storing such expressions in the data dictionary as
part of VIEW definitions or generated column definitions, which causes
wrong results.

Fixed by overriding the print() function in Item_func_like, and make
it print the ESCAPE clause if there is one.

Change-Id: I534163810b53a6ba7a4a400f7f5b3221f35675ce