Bug#32213959: PROBLEM WITH LIKE EXPRESSIONS [patch 2/3, implicit escape]
This is a preparatory refactoring to make the bug fix simpler. It
simplifies the handling of an implicit ESCAPE clause in a LIKE
expression. Instead of creating an Item_string representing the
implicit ESCAPE clause during itemize(), the escape character is set
directly to the default escape character in eval_escape_clause() when
no ESCAPE item is given.

This saves a round-trip via Item_string in the common case where no
explicit escape character is given. And it also avoids the need for
reallocating the args array in itemize() when the next patch moves
escape_item into Item_func::args.

Change-Id: I0ba3ad153b39cf122295679e39b676e3f403a853