WL#14194 step 3.3: Aliases for command line options
WL#14194: Replace old terms in replication system variables, options,
and strings

For all renamed command-line options, we re-introduce the old name as
a deprecated alias.  This includes command-line options in mysqladmin,
mysqlbinlog, mysqldump, and the server.  For the server, it only
includes command-line options that are not defined through a system
variable.

RB:25845