WL#14194 step 3.8: Aliases for system variables
WL#14194: Replace old terms in replication system variables, options,
and strings

We previously renamed several system variables.  Now we introduce a
framework for defining aliases for system variables, and define
deprecated aliases for all the renamed system variabes.

The framework consists of a subclass of Sys_var called
Sys_var_deprecated_alias. This class represents an alias for some
other Sys_var.  It copies most member fields of that variable, so
setting one of the variables changes the value of both.  We also make
it so that SET PERSIST writes the name of both variables.

When loading persisted variables, when either the alias or the real
name appears in the file, it loads both of them into the memory cache.
This ensures that both the real variable and the alias will be
persisted the next time the persisted variable file is written, which
ensures backward and forward compatibility in future upgrade/downgrade
scenarios.

In order to load the aliases while loading persisted variables, we
need to access the system variable framework (which contains the alias
definitions) while loading persisted variables.  To do that, we have
to move the initialization of the system variable framework a bit
earlier - hence the change in mysqld.cc.  This is only the
initialization of the framework, which makes it possible to access
metadata about the variable definitions; it is not the initialization
of system variable values.  The system variable framework
initialization is self-contained and does not depend on other things,
so this change should not have any other side effects.

When using RESET PERSIST, we now reset both the variable and the
alias.

When handling command-line options, we make the old name generate a
deprecation warning.  This is done in mysqld.cc:get_one_option.

We need changes to make the "variable_source" information accurate
when using an alias on the command line. (The variable_source is used
to populate columns of the performance_schema.variables_info table.)
When an alias is used on the command line, the command-line parsing
framework can only store the variable_source for the alias, since it
does not know that the alias is an alias.  Therefore, we copy the
variable_source from the alias to the base variable from code that has
this information; namely, in mysqld.cc:get_one_option.

We remove the function clear_user_host_timestamp since it was unused,
and otherwise we would have to define it in Sys_var_alias.

We make sys_var::set_and_truncate return early if source and
destination are identical. Otherwise Valgrind complains. This happens
in the new call to sysvar->set_source_name from
mysqld.cc:mysqld_get_one_option.

RB:25849