WL #14325: Interesting orders for hypergraph join optimizer [patch 6/22, collapse]
Collapse homogenized orderings into each other when it comes to
interestingness; ie., states that follow one is no better than those
that follow the other. We do this by not tracking them as interesting
orders in their own right (in the state bitmap), but instead tracking
what actually-interesting orders each state can lead to. E.g., if we
have an interesting order O_1 which gets homogenized to O_2 and O_3,
we'd used to track whether each state followed O_2 or O_3 (or both).
Instead, we now track “could become O_1” for both, which allows those
states to compare equal.

The motivating example is something like

  SELECT * FROM t1 LEFT JOIN ( SELECT * FROM t2, t3, t4 ... WHERE
     t2.x = t3.x AND t3.x = t4.x AND ... ) d ON t1.x=d.x ORDER BY t1.x;

which could cause a large number of homogenized orderings to be created,
causing access path explosion for no good reason.

Change-Id: Ib007a81cc19d7b18472d236110d5b508fcbdf15a