Bug #31613962 - NDB_RPL_DD_ADVANCE FAILS ON PB2
The test intends to sync the replication master and the slave and then
clean the slave database. However, it does not wait explicitly until
the slave is finished sync'ing, but stops the slave. Some times, this
causes the ndb object to be deleted prematurely while sync'ing is
ongoing, resulting into an assertion `m_used_cnt == 0' to fail.

Fix : Wait until slave finishes the sync'ing.

Reviewed by: Martin Skold <Martin.Skold@oracle.com>