WL #14325: Interesting orders for hypergraph join optimizer [patch 18/22, group cover]
When having interesting groupings, try to create “cover orderings” between
them and interesting orders, ie., orderings that can satisfy both.
This is so that sort-ahead for grouping can potentially satisfy
more orderings later.

Earlier, we'd simply implement group-ahead as an ordering, e.g.
for the group {a, b} we'd filesort a ASC, b ASC -- but we wouldn't
be tracking that anywhere, we'd just track the grouping, so if a
later ORDER BY would need the same ordering, we'd just re-sort.
Instead, we never produce groupings directly in the FSM; we always
create a homogenized ordering (a,b). Furthermore, we try to combine
such producing groupings with existing interesting orderings, so as
to cover those other orderings for free. For instance, if there is
a grouping {a,b,c} and an interesting ordering (c↓,a) (ie., a subset
of the grouping), we would homogenize the grouping to (c↓,a,b)
instead of (a,b,c).

This is exactly the same as we've been doing with the sorts for
DISTINCT and ORDER BY previously, but is much more generic -- in
an extreme situation, we could now (or more precisely, after the
next DISTINCT patch) do a single sort-ahead and track that ordering
throughout the tree to satisfy a semijoin grouping, GROUP BY,
DISTINCT and ORDER BY all in one with no further sorts (only some
deduplications).

Change-Id: I9a96778abfc3753d05b653cbb49cd6dbcac387ef