WL #14325: Interesting orders for hypergraph join optimizer [patch 17/22, elide temp table]
When having sorts for both GROUP BY and ORDER BY, we insert a temporary
table in the middle, because it is needed when aggregates are part of
the sort. However, this poses a problem for unit tests, which don't
initialize enough of the server to instantiate temporary tables.

In a few select cases, notably when we don't actually compute any
aggregates and when there are no row IDs needed from the base tables,
we can elide this temporary table. Do so to make it possible to unit
test a GROUP BY + ORDER BY combination in the next patch.

There is also a performance benefit in a few select cases, but it's
so small that it is not the main point of the patch.

Change-Id: Ia0f8cab00ff057959f1479e66ade9b450be39320