Bug#32392898 - WRONG VIEW-ID WRITTEN TO STATEFILE FAILOVER WITH ASYNC REPLICASETS
Problem
=======
WL#13327 introduced a mechanism for stopping and starting a listening
socket acceptors. In router using metadata cache scenario if the acceptor
could not be started another attempt is made on the next metadata refresh.
That was done by using the 'instances has changed mechanism', which caused
the statefile update without a view-id.

Fix
===
1. Rename the metadata cache state listener
2. Refactor the start/stop socket mechanism and extract the start/stop
   methods as callbacks. That allowed to remove the 'wait for socket state
   notification' loop from the router plugin.
3. Add new callback that will be called when we need to update the socket
   acceptor state when doing a metadata refresh.

RB: 25886
Reviewed by: Jan Kneschke <jan.kneschke@oracle.com>