Bug #32015466: MYSQL BATCH MODE EXECUTE SQL FAILED WITH MYSQL CLIENT
Description:
When 'use' statement with "`" is after a "DROP DATABASE" statement
there is an ERROR saying "Unknown database".

Analysis:
In com_go(), get_current_db() is not called in batch mode. So the
'current_db' is not updated after a database is dropped.

Fix:
Calling get_current_db() in batch mode ensures that the
'current_db' is updated.

RB#25703