Bug 32367250 - ABORTED NDB BACKUP DUE TO FULL FILE SYSTEM CAUSES CLUSTER CRASH
Data nodes crash on encountering an 'out-of-disk-space' error during
a compressed backup. The crash occurs when NDBFS attempts to append
a block of data to the backup file with a compressed write and fails.
The return code which indicates an error for compressed writes is
incorrectly treated as a fatal error and a crash is triggered in
response.

This is fixed by adding a check to handle the error return code from
compressed writes. The system error is extracted and set so that the
backup fails with an appropriate error "No space left on the device:
Temporary error: Temporary Resource error".

An mtr test using error insertion is added to test this scenario. Error
inserts are implemented to simulate out-of-disk errors in ndbzwrite
for writes to backup datafiles.

Reviewed-by: Frazer Clement <frazer.clement@oracle.com>
Change-Id: Ia9d67188a95e7e14bef6a032e5b472ada954700b