Bug#32348682: WL#13520: WRONG RESULT WHEN PREDICATE IN A CORRELATED               SUBQUERY HAS EXPRESSIONS
To transform a correlated subquery to derived table,
in the correlated predicate, if the non-correlated operand is an
expression (i.e not a single column reference), the grouping
in the derived table needs to be on the expression and not
on the columns present in the expression. Otherwise, it would
lead to wrong results in certain cases.
Since the transformation from subquery to derived table, currently
supports grouping only on column references, we will be rejecting
this transformation for subqueries having such predicates.

Bug#32378012: ASSERTION `GOING.ELEMENTS >= 1' FAILED IN SQL/SQL_RESOLVER.CC

A function having a correlated field, if determined to be a
constant (in this case ISNULL(non-nullable correlated field))
will have its used_tables as 0 although the field itself
would still be marked as an outer reference.
This leads to an inconsistency in checks performed during
transformation of a subquery to derived table.
While checking for correlated fields in a predicate during
transformation, since the field is still marked as
such, it is concluded that we have a correlated predicate.
However, while extracting the correlated predicate,
since the predicate itself is not marked as having an
outer reference, it leads to an assert.

Since we reject queries having expressions in a non-correlated
operand of a correlated predicate as part of the fix for Bug#32348682,
we would be rejecting this query too as the operand (expression)
is not marked correlated.

Reviewed By: Roy Lyseng <roy.lyseng@oracle.com>