WL #14325: Interesting orders for hypergraph join optimizer [patch 1/22, Fake_TABLE]
Fake_TABLE has a global counter of how many tables it has created, which
it uses for table numbers. Reset it every time we set up the fake server
initializer, so that table numbers are consistent between merged and
non-merged unit test runs. Otherwise, we could be unlucky with
wraparound and have a situation where a unit test expects tables to be
“ordered” as t1,t2,t3 where it became t3,t1,t2 or similar instead,
due to unlucky wraparound.

Change-Id: If79ac6835a3b64a46ee73d9819b235fbeab0160a