WL #14418: Support for cyclic hypergraphs [patch 5/16, Item_equal_iterator]
Pre-patch: Replace Item_equal_iterator with the more modern STL-compatible
iterators. We add a simple proxy to avoid exposing the fields list
directly (making it easier to convert it from List<Item> to
Mem_root_array<Item *> in the future), but could probably have done
so if we wanted to.

Change-Id: Ib1f9932e997db4a7894552288ab7abf2ce2aa1e7