BUG#32433002 NDB_MGMD FAILING ON WINDOWS PB2 WITH TIMEOUT ERROR
ndb_mgmd when started on port that was already in use did not throw
any error since the use of SO_REUSEADDR allows multiple sockets to
bind to same address and port in windows.

SO_REUSEADDRPORT is replaced with SO_EXCLUSIVEADDRUSE for preventing
start of mgmd on port that is already in use. The error code for any
failures in socket has to be fetched as soon as failure occurs as the
other socket methods override the error code to 0 on success.
tryBind is modified to extract error_message before closing socket and
optionally return to the calling function.
Added a method in portlib to return error message as a string using
FormatMessage for windows and strerror() for other platforms.

Change-Id: I8cab9e491b4fe271a6db6827d044c785a15e88d6