WL #14418: Support for cyclic hypergraphs [patch 12/16, complete mesh]
When we have a multi-equality within a companion set, add full
all-to-all graph edges, so that the join optimizer is free to choose
any join order. (This is generally what's thought of as full
multi-equality support.)

Reenable the information_schema queries we disabled earlier, as the
fast plan is now available (and chosen) again.

Change-Id: I84eeeb17e4a8d8398d4f318fb89dea3f6caa990a