WL#14422: Full-text search in the hypergraph join optimizer [patch 6/12, index]
Enable use of full-text index scans when using MATCH as a WHERE
predicate. If MATCH is used a predicate, we propose both a table scan
and a full-text index scan using that predicate, and pick the one
which gives the lowest cost in the final plan.

Only calls to MATCH directly contained in the WHERE clause are
considered for full-text index scans. MATCH calls nested in other
functions in the WHERE clause are not considered. Some of these will
be considered by the next patch.

Change-Id: Ia7cc5c21a5759f44a0f57180674eee82d58b2086