Bug#32068551 NDB: IDLE FRAGMENT LCPS AFFECT REQUEST LATENCY
The Local Checkpoints (LCP) mechanism was changed in 7.6 such
that it also detected 'idle' fragments, that is fragments which
was not changes since last and only required a on-disk metadata
update. The LCP mechanism could then immediately proceed to
handle the next fragment.

With a high rate of such idle fragments, it turns out that the CPU
consumption to just loop through these becomes signifficant, causing
latency spikes in user transactions.

A 1ms delay was already inserted between each such idle fragment being handled.
However, tests has proved that to be too short, and we are normally not
in a hurry to complete these idle fragments after all.

This patch extends this idle fragment delay to 20ms if there are no
'redo alerts', which would indicate that there are some urgency to
complete the LCP. In case of a 'low redo alert' state it will wait 5ms
instead, and for a higher alert state it fall back to the 1ms delay.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>