Bug#32525732 MTR TESTS ARE SIGNIFICANTLY SLOWER ON WINDOWS THAN LINUX
Added a new CMake option WIN_DEBUG_RTC (defaulted to OFF) that controls
the addition of the MSVC /RTC1 compiler flag when creating debug builds.

/RTC1 checks slow down MTR testing significantly on Windows but rarely
catch bugs.  The checks that are not performed when /RTC1 is not used on
Windows are still performed by Linux ASAN and valgrind tests, and also
detected by compiler warnings.

RB: 25960