Bug#32529380 routertest_component_socket_close is slow
Problem
=======

routertest_component_socket_close takes 210 secs to complete (on Linux).

The tests spend ~18 secs checking if a socket is in use by another
application (socket-blocker).

The socket-blocker does:

- bind()
- listen()

The tester does:

- connect()

As the connect() connects to a socket that is bound, it will wait until
it is accept()ed ... or timeout (after [system-dependent] 18sec).

Change
======

- lower metadata-cache TTL to 100ms
- accept() the connect() and send a MySQL Error to fails the
  mysql_real_connect().

Test runtime goes from 210sec to 44sec.

RB: 25963