Bug#32448364 "TABLE IN DB WITH UPPERCASE NAME IS NOT SHOWN IN I_S IF LOWER_CASE_TABLE_NAMES=2".
On systems running in --lower-case-table-names=2 mode privilege
checks used in implementation of some I_S tables (I_S.TABLES,
I_S.ROUTINES and so on) didn't work correctly for databases and
tables with uppercase in their names for some users. For example,
this led to tables in database with uppercase in its name not
shown in I_S.TABLES to users which had table-level privileges
on these tables.

The problem was caused by the fact that functions implementing
these privilege checks, e.g. CAN_ACCESS_TABLE(), didn't take
into account the fact that in --l-c-t-n=2 mode database names,
which are passed to them as arguments, have their original case
(since this is how they are stored in the data-dictionary),
while ACL calls used by these functions assume that database
and tables names are in lower case if l-c-t-n > 0.

The same problem for table names was earlier addressed by fix
for bug 29957361. Unfortunately, the latter fix missed case of
CAN_ACCESS_COLUMN().

This patch fixes the problem by changing implementation of
CAN_ACCESS_TABLE() and other similar functions to lowercase
both database and table names in l-c-t-n=2 mode before passing
them to ACL calls.

Reviewed-by: Gopal Shankar <Gopal.Shankar@oracle.com>