BUG 32529561 - CRASH IF ROLLBACK A TRANSACTION THAT MODIFIES GENERATED AND INDEXED COLUMN
The crash happens while attempting to free up the space occupied by the
externally stored columns.
The update vector containing the columns which consume the external storage
space does not include the virtual generated columns.
While iterating through the columns to find out which columns need to free up
the space, the server hit the assertion as the column being freed up is a
virtually generated one and is not accounted for in the update vector.

Solution
--------
Check if the column is a virtual column before checking if it is part of the update
vector.

RB# 25993
Reviewed By: Annamalai Gurusami <annamalai.gurusami@oracle.com>