Bug#32356419 USE NODE.JS 12.20.1 IN NDB
Update NDB's node.js code to use Node.js version 12.20.1

The ReadMe file in storage/ndb/nodejs was unclear about the steps
needed to test the node.js adapter. Edit the file to make it clear
that testing can be run from a mysql install directory (but not
a mysql source directory or build directory) and to better describe
the needed steps.

Delete the README.md file installed in share/nodejs/jones-ndb as
this distracts from the better ReadMe file one directory up.

Change the directory structure and the jones-ndb/package.json file
to work better with the current version of npm.

Add a create-test-user.sql script to compensate for the incompatibility
between the node-mysql module (used by jones-mysql) and MySQL server
8.0 authentication.

The fix in AsyncMethodCall.h eliminates a compiler warning. The fixes
in Record.cpp and JsWrapper.h allow cast-qual to be removed from the
list of ignored warnings.

Approved-by: Kari Juul Wedde <kari.wedde@oracle.com>