Bug#29363867: LOST CONNECTION TO MYSQL SERVER DURING QUERY
Note: this patch is for 5.7.

This is based on original patch and comments of g.manasa@oracle.com

Description: Server exits when a prepared statement is executed after uninstalling a plugin.

Analysis: As per the current design, there is no error checking or locking for plugin variables used in EXECUTE statement. When EXECUTE command is run after uninstalling the plugin, server exits with a segmentation fault since the plugin variable is freed. Even for concurrent EXECUTE and UNINSTALL, the server crashes with a segmentation fault.

Fix: All plugin variables are rebound in prepared statements before accessing their data.

Change-Id: I0db3702f579ba470b53a5b21802c6435f11b23f6