Bug #32555134 ADD SUPPORT FOR CLANG 11 ON SOLARIS [noclose]
Workaround for bug in the standard library (or clang *use* of the
standard library) We get Undefined symbol for some template
implementing std::time_get and std::time_put.

The workaround is to use the C functions strftime and strptime instead.
strptime() is not available on Windows, so add a feature check for it,
and fallback to the <iomanip> implementation.

For Solaris we set the cmake variable LIBM unconditionally (math library is needed).
Do not overwrite or reset it in configure.cmake

Update the platform ifdefs in mysys/my_rdtsc.cc so that we pick a
solution which works for clang on Solaris. Remove tests for
__SunOS_5_7.  Clang will set the C macro __GNUC__, but does not handle
the same __asm directives as gcc.

Adjust some cmake code which tests for SOLARIS and GNU/CLANG compiler,
so we get a clean build.

Fix some -Wsign-compare warnings.

With this patch, the *build* is clean. There are however many failing
tests.  There seems to be a bug in atexit() processing, so all static
objects in shared libraries must be trivially destructible (which is
good programming practice anyways...)
This affects several plugins, in the server and for router.

Clang seems to generate better code on SPARC (faster and smaller) than
gcc and Developer Studio. Sysbench TPS is up ~20%.

Change-Id: Ifc2d6436686af42d103f8c2f8ceb3248e2270f19